Made notes on the direction for some of the things I think will be needed and how I think they could emerge.
The notes are not typically included in source control unless they are needed to capture
the information for sharing with the team or for audit purposes.  For either of these situations
I would suspect that the notes would be more formally structured.

The purpose of these notes is to capture the thought process during development, I feel that loose structure is ok.