Added Comments capturing the rules to be implemented.
Came accross an interesting issue and made notes of it.

If deciding to use mocks for testing is something that almost seems
to be the default. Sometimes people need to think harder about it.

In this case I decide to try to avoid mock as we can proceed without
needing it.  Will be interesting to see if this comes back to haunt me.