Added Rules class.
While thinking about starting the grid it seemed like having something
to calculate the next gen status of a cell would be needed fairly early so
figured geting the rules in would be a good next step.

The first step when starting a new class is to make sure the test wiring is good.
I wouldn't typically commit a wiring test as it live for only a short while.