Added a setup function for use in testing the rules class.
Discovered that it is looking like mock may be the way to go.
Still exploring non-mocking to carry out the thought experiment and
see if the cell class implementation will be changed and cause this test to
explode.

If time permits I may pretend that there is a business reason to change the
cell class in a way that will break this.  Mostly to see how far reaching
the breakage is.