Red->Green->Refactor is complete.
This is where a normal commit might happen, might wait until the next test is done,
depends on if it is lunchtime or not.  Always commit before lunch (wiyghbat)
Message:

Added logic for dead cells that should stay dead.

When a cell is dead and the count of neighbors is not three is should stay dead.