Added test for dead cell that should become alive.
No code change was needed as the code was already in a state allowing this
test to pass.  The test is needed to keep the behaviour pinned in place.

If the previous tests were made to pass in a different way it is possible
that this test could have caused a red-green-refactor cycle.