Added a test that tests all the things via test cases in a single parameterized test.
I really do not like this as it globs the conditions together and becomes
less clear about what the intent is.

Also it is difficult to name the test, which is a clue that it is smelly.

Thought it would be interesting to try, normally, this would be either a first
attempt and broken up or if it was thought about later, would be deleted without
hitting source control.  Thought it was interesting though so included it here
to think about later.