Clean up rules tests.
Removed comments that are no longer needed.
Decided to leave the setup as is as.  It is test code which is not an excuse
to intentionally make it dirty.  At the moment I do not see a way to clean
it up.  Likely I would make a note of this and check it the next day to
see if any bright ideas pop into my head.

I think part of the reason I am stuck on it is that I am anxious to get
to the grid code and get something on the screen.

Stay tuned, I think we are about to discover a limitation on the cell class,
there is not currently a provision to update it without botching up the status
for neighbors to look at while updating the grid.
I suspect we will end up with a flag on the cell indicating it's next status
and also suspect we will need to scan the grid to check for updates, then
run the grid again to effect the update. This implementation is turning out
to be a bit heavy. (lots of rather bigish cell objects and running the grid
twice per update)  Not worried about that though, this is the first go round
and I think it will be more than fast enough to keep the business people
happy while we figure out how to rework things.  Also it may be good forever
I do not recall having seen any perf spec.