Decided grid builder should return a root cell instead of a grid.
It is possibly Grid.cs may go away entirely. Will be interesting to see
how this develops.

Current thinking is that we will have a builder, then a traverser. Not sure
how the view will turn out yet but is not difficult to envision that it
will be yet another run of the traverser.